% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{aggr}
\alias{aggr}
\title{Generic for Accessing Aggregated Predictions}
\usage{
aggr(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Aggregated predictions (format depends on the object class).
}
\description{
Dispatches to methods for extracting aggregated predictions from objects.
}
