% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{coef.bgnlm_model}
\alias{coef.bgnlm_model}
\title{Coefficients for BGNLM Model}
\usage{
\method{coef}{bgnlm_model}(object, ...)
}
\arguments{
\item{object}{Object of class "bgnlm_model".}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of coefficients.
}
\description{
Extracts coefficients from a BGNLM model.
}
\examples{
data(exoplanet)
model <- get.best.model(fbms(semimajoraxis ~ ., data = exoplanet, family = "gaussian"))
coef(model)
}
