% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.fbms_predict}
\alias{fitted.fbms_predict}
\title{Access Fitted Values}
\usage{
fitted.fbms_predict(object, ...)
}
\arguments{
\item{object}{Object of class "fbms_predict".}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of mean predictions.
}
\description{
Extracts the mean predictions from an FBMS prediction object (alias for mean).
}
\examples{
\donttest{
data(exoplanet)
model <- fbms(semimajoraxis ~ ., data = exoplanet)
pred <- predict(model, exoplanet[51:60, -1])
fitted(pred)
}
}
