% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{glm.logpost.bas}
\alias{glm.logpost.bas}
\title{Log likelihood function for glm regression with parameter priors from BAS package}
\usage{
glm.logpost.bas(y, x, model, complex, mlpost_params = NULL)
}
\arguments{
\item{y}{A vector containing the dependent variable}

\item{x}{The matrix containing the precalculated features}

\item{model}{A logical vector indicating which features are included in the model}

\item{complex}{A list of complexity measures for the features}

\item{mlpost_params}{A list of parameters for the log likelihood, supplied by the user}
}
\value{
A list with the approximate log marginal likelihood combined with the log prior (\code{crit}) and the posterior mode of coefficients (\code{coefs})
}
\description{
This is a placeholder version of the function.
It falls back to \code{glm.loglik.g} and raises a warning if the full function is not loaded.
}
\examples{
glm.logpost.bas(as.integer(rnorm(100) > 0), 
cbind(1, matrix(rnorm(100))), c(TRUE, TRUE), 
list(oc = 1))

}
