% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{nrelu}
\alias{nrelu}
\title{Negative ReLU Function}
\usage{
nrelu(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
max(-x,0)
}
\description{
Negative ReLU Function
}
\examples{
nrelu(2)

}
