% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{p0}
\alias{p0}
\title{p0 Polynomial Term}
\usage{
p0(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
log(abs(x) + .Machine$double.eps)
}
\description{
p0 Polynomial Term
}
\examples{
p0(2)

}
