% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{p05}
\alias{p05}
\title{p05 Polynomial Term}
\usage{
p05(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
(abs(x)+.Machine$double.eps)^(0.5)
}
\description{
p05 Polynomial Term
}
\examples{
p05(2)

}
