% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.bgnlm_model}
\alias{plot.bgnlm_model}
\title{Plot BGNLM Model}
\usage{
\method{plot}{bgnlm_model}(x, ...)
}
\arguments{
\item{x}{Object of class "bgnlm_model".}

\item{...}{Additional arguments passed to barplot.}
}
\value{
The input object (invisibly).
}
\description{
Plots the coefficients of a BGNLM model.
}
\examples{
data(exoplanet)
model <- get.best.model(fbms(semimajoraxis ~ ., data = exoplanet, family = "gaussian"))
plot(model)
}
