% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.gmjmcmc}
\alias{plot.gmjmcmc}
\title{Function to Plot GMJMCMC Results and Merged Results from merge.results}
\usage{
\method{plot}{gmjmcmc}(x, count = "all", pop = "best", tol = 1e-07, data = NULL, ...)
}
\arguments{
\item{x}{The results to use}

\item{count}{The number of features to plot, defaults to all}

\item{pop}{The population to plot, can be "last", "best" or "highest_mass". Defaults to "best".}

\item{tol}{The tolerance to use for the correlation when finding equivalent features, default is 0.0000001}

\item{data}{Data to merge on, important if pre-filtering was used}

\item{...}{Not used.}
}
\value{
No return value, just creates a plot
}
\description{
Function to Plot GMJMCMC Results and Merged Results from merge.results
}
\examples{
result <- gmjmcmc(
  y = matrix(rnorm(100), 100),
  x = matrix(rnorm(600), 100),
  P = 2,
  transforms = c("p0", "exp_dbl")
)
plot(result)

}
