% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.mjmcmc_parallel}
\alias{plot.mjmcmc_parallel}
\title{Plot an mjmcmc_parallel Run}
\usage{
\method{plot}{mjmcmc_parallel}(x, count = "all", ...)
}
\arguments{
\item{x}{The results to use}

\item{count}{The number of features to plot, defaults to all}

\item{...}{Not used.}
}
\value{
No return value, just creates a plot
}
\description{
Plot an mjmcmc_parallel Run
}
\examples{
result <- mjmcmc.parallel(
  runs = 1,
  cores = 1,
  y = matrix(rnorm(100), 100),
  x = matrix(rnorm(600), 100),
  loglik.pi = gaussian.loglik
)
plot(result)

}
