% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predmean.fbms_predict}
\alias{predmean.fbms_predict}
\title{Access Mean Predictions}
\usage{
\method{predmean}{fbms_predict}(object, ...)
}
\arguments{
\item{object}{Object of class "fbms_predict".}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of mean predictions.
}
\description{
Extracts the mean predictions from an FBMS prediction object.
}
\examples{
\donttest{
data(exoplanet)
model <- fbms(semimajoraxis ~ ., data = exoplanet, method = "mjmcmc")
pred <- predict(model, exoplanet[51:60, -1])
predmean(pred)
}
}
