% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predmean}
\alias{predmean}
\title{Generic for Accessing Mean Predictions}
\usage{
predmean(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Posterior mean predictions (format depends on the object class).
}
\description{
Dispatches to methods for extracting quantile predictions from objects.
}
