% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predquantiles}
\alias{predquantiles}
\title{Generic for Accessing Quantile Predictions}
\usage{
predquantiles(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Quantile predictions (format depends on the object class).
}
\description{
Dispatches to methods for extracting quantile predictions from objects.
}
