% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bgnlm_model}
\alias{print.bgnlm_model}
\title{Print BGNLM Model Object}
\usage{
\method{print}{bgnlm_model}(x, ...)
}
\arguments{
\item{x}{Object of class "bgnlm_model".}

\item{...}{Additional arguments (ignored).}
}
\value{
Prints a summary of the model and returns NULL
}
\description{
Displays the coefficients of a BGNLM model object.
}
\examples{
data(exoplanet)
model <- get.best.model(fbms(semimajoraxis ~ ., data = exoplanet, 
family = "gaussian"))
print(model)
model <- get.mpm.model(fbms(semimajoraxis ~ ., data = exoplanet, 
family = "gaussian"), y = exoplanet[,1],x = exoplanet[,-1])
print(model)
}
