% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{print.fbms_predict}
\alias{print.fbms_predict}
\title{Print FBMS Prediction Object}
\usage{
\method{print}{fbms_predict}(x, ...)
}
\arguments{
\item{x}{Object of class "fbms_predict".}

\item{...}{Additional arguments (ignored).}
}
\value{
Prints a summary and returns NULL.
}
\description{
Displays a summary of an FBMS prediction object, including mean predictions and quantiles.
}
\examples{
\donttest{
data(exoplanet)
model <- fbms(semimajoraxis ~ ., data = exoplanet)
pred <- predict(model, exoplanet[51:60, -1])
print(pred)
}
}
