% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gmjmcmc}
\alias{print.gmjmcmc}
\title{Print GMJMCMC Model Object}
\usage{
\method{print}{gmjmcmc}(x, ...)
}
\arguments{
\item{x}{Object of class "gmjmcmc".}

\item{...}{Additional arguments passed to summary method.}
}
\value{
Prints a summary of the model and returns NULL
}
\description{
Displays a concise summary of a GMJMCMC model object.
}
\examples{
data(exoplanet)
model <- fbms(semimajoraxis ~ ., data = exoplanet,method = "gmjmcmc", transforms = c("sigmoid"))
print(model)
}
