% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.bgnlm_model}
\alias{residuals.bgnlm_model}
\title{Residuals for BGNLM Model}
\usage{
\method{residuals}{bgnlm_model}(object, y, x, ...)
}
\arguments{
\item{object}{Object of class "bgnlm_model".}

\item{y}{Respnse.}

\item{x}{Covariates.}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of residuals.
}
\description{
Computes residuals as the difference between observed and predicted values.
}
\examples{
library(FBMS)
data(exoplanet)
model <- get.best.model(fbms(semimajoraxis ~ ., data = exoplanet, family = "gaussian"))
hist(residuals(model, exoplanet[,1], exoplanet[,-1]))
}
