% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.gmjmcmc}
\alias{residuals.gmjmcmc}
\title{Residuals for GMJMCMC Model}
\usage{
\method{residuals}{gmjmcmc}(object, y, x, ...)
}
\arguments{
\item{object}{Object of class "gmjmcmc".}

\item{y}{Respnse.}

\item{x}{Covariates.}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of residuals.
}
\description{
Computes residuals as the difference between observed and predicted values.
}
\examples{
data(exoplanet)
model <- fbms(semimajoraxis ~ ., data = exoplanet, method = "gmjmcmc", transforms = c("sigmoid"))
hist(residuals(model, exoplanet[,1], exoplanet[,-1]))
}
