% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.mjmcmc_parallel}
\alias{residuals.mjmcmc_parallel}
\title{Residuals for MJMCMC Parallel Model}
\usage{
\method{residuals}{mjmcmc_parallel}(object, y, x, ...)
}
\arguments{
\item{object}{Object of class "mjmcmc_parallel".}

\item{y}{Respnse.}

\item{x}{Covariates.}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of residuals.
}
\description{
Computes residuals as the difference between observed and predicted values.
}
\examples{
data(exoplanet)
model <- fbms(semimajoraxis ~ ., data = exoplanet, method = "mjmcmc.parallel",runs = 2, cores = 1)
hist(residuals(model, exoplanet[,1], exoplanet[,-1]))
}
