% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{string.population}
\alias{string.population}
\title{Function to Get a Character Representation of a List of Features}
\usage{
string.population(x, round = 2)
}
\arguments{
\item{x}{A list of feature objects}

\item{round}{Rounding precision for parameters of the features}
}
\value{
A matrix of character representations of the features of a model.
}
\description{
Function to Get a Character Representation of a List of Features
}
\examples{
result <- gmjmcmc(
  y = matrix(rnorm(100), 100),
  x = matrix(rnorm(600), 100),
  P = 2,
  transforms = c("p0", "exp_dbl")
)
string.population(result$populations[[1]])

}
