% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{troot}
\alias{troot}
\title{Cube Root Function}
\usage{
troot(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
The cube root of x
}
\description{
Cube Root Function
}
\examples{
troot(27)

}
