% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.mjmcmc}
\alias{plot.mjmcmc}
\title{Function to Plot GMJMCMC Results and Merged Results from merge.results}
\usage{
\method{plot}{mjmcmc}(x, count = "all", ...)
}
\arguments{
\item{x}{The results to use}

\item{count}{The number of features to plot, defaults to all}

\item{...}{Not used.}
}
\value{
No return value, just creates a plot
}
\description{
Function to Plot GMJMCMC Results and Merged Results from merge.results
}
\examples{
result <- mjmcmc(
y = matrix(rnorm(100), 100),
x = matrix(rnorm(600), 100),
loglik.pi = gaussian.loglik)
plot(result)

}
